refactor(beacon): move ensure_beacon_dir to core (PER-120)
ensure_beacon_dir is consumed by both the setup and warehouse domains,
making it a cross-domain primitive. Move it to beacon.core.beacon_dir
and update connector.py to import from there, eliminating the
warehouse→setup cross-domain coupling.
