+++ libs/beacon/src/beacon/domains/setup/initializer.py
@@ -232,13 +232,3 @@ class WarehouseInitializer:
             raise
         except FileNotFoundError:
             logger.warning("Git not found in PATH, skipping git initialization")
-
-
-def ensure_beacon_dir(project_root: Path) -> Path:
-    """Ensure .agentic-beacon directory exists, creating it if necessary.
-
-    Returns the path to the beacon directory.
-    """
-    beacon_dir = project_root / ".agentic-beacon"
-    beacon_dir.mkdir(exist_ok=True)
-    return beacon_dir
